Change to integer instead of float to get rid of unsightly decimal

+++ auctions/models.py
@@ -12,7 +12,7 @@ class Category(models.Model):
       return self.categoryName
 
 class Bid(models.Model):
-   bid = models.FloatField(default=0)
+   bid = models.IntegerField(default=0)
    user = models.ForeignKey(User, on_delete=models.CASCADE, blank=True, null=True, related_name="userBid")
 
 class Listing(models.Model):
